app/requirements.txt
transformers
Pillow
requests
fastapi
uvicorn
logging